Add recipes folder and fix links on recipe pages
I have added the directory for recipies and added the .html file for
each recipe with in this. This also broke the links to the images
folder and for the recipe pages link to index.html and recipes.html.
To resolve this I have change the relative links to go up one more
directory but changing "./" to "../"